utils, queryの配置の見直し
